requirements.txt
crewai>=0.11.0
streamlit>=1.32.0
google-generativeai>=0.3.0
lottie>=0.7.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
streamlit-lottie>=0.0.3 